+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/Clouds.java
@@ -1,0 +1,4 @@
+package com.rperazzo.weatherapp.model;
+
+public class Clouds {
+}
