+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/presenter/MainActivityPresenterImpl.java
@@ -1,0 +1,4 @@
+package com.rperazzo.weatherapp.presenter;
+
+public class MainActivityPresenterImpl implements MainActivityPresenter {
+}
