+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/view/MainActivity.java
@@ -1,0 +1,146 @@
+package com.rperazzo.weatherapp.view;
+
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.net.ConnectivityManager;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.KeyEvent;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.view.inputmethod.InputMethodManager;
+import android.widget.EditText;
+import android.widget.ListView;
+import android.widget.ProgressBar;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import com.rperazzo.weatherapp.R;
+import com.rperazzo.weatherapp.model.City;
+import com.rperazzo.weatherapp.presenter.MainActivityPresenter;
+import com.rperazzo.weatherapp.presenter.MainActivityPresenterImpl;
+import com.rperazzo.weatherapp.repository.WeatherManager;
+import com.rperazzo.weatherapp.repository.WeatherManager.FindResult;
+import com.rperazzo.weatherapp.view.adapter.FindItemAdapter;
+
+import java.util.ArrayList;
+
+public class MainActivity extends AppCompatActivity implements MainActivityView {
+
+    private EditText mEditText;
+    private TextView mTextView;
+    private ProgressBar mProgressBar;
+    private ListView mList;
+    private FindItemAdapter mAdapter;
+    private MainActivityPresenter presenter;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        initViews();
+
+        presenter = new MainActivityPresenterImpl(this);
+        presenter.onCreate();
+        presenter.initializeLocationModel(this);
+
+        mEditText.setOnKeyListener(new View.OnKeyListener() {
+            public boolean onKey(View v, int keyCode, KeyEvent event) {
+                if ((event.getAction() == KeyEvent.ACTION_DOWN) &&
+                        (keyCode == KeyEvent.KEYCODE_ENTER)) {
+                    presenter.searchByName();
+                }
+                return false;
+            }
+        });
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.main_menu, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        int id = item.getItemId();
+
+        if (id == R.id.menu_celcius) {
+            presenter.updateUnitIfNecessary("metric");
+            return true;
+        } else if (id == R.id.menu_fahrenheit) {
+            presenter.updateUnitIfNecessary("imperial");
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+
+    @Override
+    public ConnectivityManager returnSystemService() {
+        return (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+    }
+
+    @Override
+    public void showToast(String message) {
+        Toast.makeText(this, message, Toast.LENGTH_LONG).show();
+    }
+
+    @Override
+    public String getSearch() {
+        return mEditText.getText().toString();
+    }
+
+    @Override
+    public void onStartLoading() {
+        mList.setVisibility(View.GONE);
+        mProgressBar.setVisibility(View.VISIBLE);
+        mTextView.setVisibility(View.GONE);
+
+        View view = this.getCurrentFocus();
+        if (view != null) {
+            InputMethodManager imm = (InputMethodManager)
+                    getSystemService(Context.INPUT_METHOD_SERVICE);
+            imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
+        }
+    }
+
+    @Override
+    public void onFinishLoading(FindResult result) {
+        mProgressBar.setVisibility(View.GONE);
+        mList.setVisibility(View.VISIBLE);
+        mAdapter.notifyDataSetChanged();
+    }
+
+    @Override
+    public void onFinishLoadingWithoutResults() {
+        mTextView.setText("No results.");
+    }
+
+    @Override
+    public void onFinishLoadingWithError() {
+        mProgressBar.setVisibility(View.GONE);
+        mList.setVisibility(View.GONE);
+        mTextView.setText("Error");
+    }
+
+
+    @Override
+    public void setupAdapter(ArrayList<City> cities) {
+        mAdapter = new FindItemAdapter(this, cities);
+        mList.setAdapter(mAdapter);
+    }
+
+    public void onSearchClick(View view) {
+        presenter.searchByName();
+    }
+
+    private void initViews() {
+        mEditText = (EditText) findViewById(R.id.editText);
+        mTextView = (TextView) findViewById(R.id.textView);
+        mProgressBar = (ProgressBar) findViewById(R.id.progressBar);
+        mList = (ListView) findViewById(R.id.list);
+    }
+}
