+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/view/MainActivityView.java
@@ -1,0 +1,4 @@
+package com.rperazzo.weatherapp;
+
+public interface MainActivityView {
+}
